Quick-Save somewhat implemented. Had to mess with a lot of files to get this feature implemented so there are still A LOT of bugs but it shouldn't crash on start. If there are any merging conflicts with what any of you are working on, my bad.
Readded Save & Quit button to the main scene so the player can save mid game and come back later. If a quick save file is found on the title screen, the continue button will show up. 

Players can also just click New to start a new game. 

Added a saving folder to the editor. Made a lot of previously private variables public. in basically all the scripts that got touched to help with the testing. Probably fix that at a later date.